requirements.txt
streamlit==1.32.0
pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
scikit-learn==1.3.1
matplotlib==3.8.0 